requirements.txt
chromadb==0.4.4
langchain==0.0.334
openai
transformers==4.30.2
pypdf
streamlit
streamlit_chat
sentence_transformers
tiktoken
langchain_experimental==0.0.35
faiss-cpu
